smartdashboard add
-reset gyros

+++ src/main/java/frc/robot/subsystems/Intake.java
@@ -15,10 +15,10 @@ public class Intake extends SubsystemBase {
 
   public Intake(int intakePort, int troughPort) {
     intakeMotor = new CANSparkMax(intakePort, MotorType.kBrushed);
-    troughMotor = new CANSparkMax(troughPort,MotorType.kBrushed);
+    troughMotor = new CANSparkMax(troughPort, MotorType.kBrushed);
   }
 
-  public void set(float speed) {
+  public void set(double speed) {
     intakeMotor.set(speed);
     troughMotor.set(speed);
   }
